feat(error handling): refactor ExpandCollapse
